+++ controllers/hvpa_controller.go
@@ -29,7 +29,6 @@ import (
 
 	autoscalingv1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
-	autoscalingv1 "k8s.io/api/autoscaling/v1"
 	autoscaling "k8s.io/api/autoscaling/v2beta2"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -275,51 +274,6 @@ func (r *HvpaReconciler) deleteHvpaFinalizers(hvpa *autoscalingv1alpha1.Hvpa) {
 	}
 }
 
-func getVpaFromHvpa(hvpa *autoscalingv1alpha1.Hvpa) (*vpa_api.VerticalPodAutoscaler, error) {
-	metadata := hvpa.Spec.Vpa.Template.ObjectMeta.DeepCopy()
-
-	labels := metadata.GetLabels()
-	if labels == nil || len(labels) == 0 {
-		// TODO: Could be done better as part of validation
-		return nil, fmt.Errorf("Need labels in VPA template")
-	}
-
-	if ownerRef := metadata.GetOwnerReferences(); len(ownerRef) != 0 {
-		// TODO: Could be done better as part of validation
-		return nil, fmt.Errorf("vpa template in hvpa object already has an owner reference")
-	}
-
-	if generateName := metadata.GetGenerateName(); len(generateName) != 0 {
-		log.V(3).Info("Warning", "Generate name provided in the vpa template will be ignored", generateName)
-	}
-
-	if name := metadata.GetName(); len(name) != 0 {
-		log.V(3).Info("Warning", "Name provided in the hpa template will be ignored", name)
-		metadata.SetName("")
-	}
-
-	metadata.SetGenerateName(hvpa.Name + "-")
-	metadata.SetNamespace(hvpa.Namespace)
-
-	// Updater policy set to "Off", as we don't want vpa-updater to act on recommendations
-	updatePolicy := vpa_api.UpdateModeOff
-
-	return &vpa_api.VerticalPodAutoscaler{
-		ObjectMeta: *metadata,
-		Spec: vpa_api.VerticalPodAutoscalerSpec{
-			TargetRef: &autoscalingv1.CrossVersionObjectReference{
-				Name:       hvpa.Spec.TargetRef.Name,
-				APIVersion: hvpa.Spec.TargetRef.APIVersion,
-				Kind:       hvpa.Spec.TargetRef.Kind,
-			},
-			ResourcePolicy: hvpa.Spec.Vpa.Template.Spec.ResourcePolicy.DeepCopy(),
-			UpdatePolicy: &vpa_api.PodUpdatePolicy{
-				UpdateMode: &updatePolicy,
-			},
-		},
-	}, nil
-}
-
 func (r *HvpaReconciler) reconcileVpa(hvpa *autoscalingv1alpha1.Hvpa) (*vpa_api.VerticalPodAutoscalerStatus, error) {
 	selector, err := metav1.LabelSelectorAsSelector(hvpa.Spec.Vpa.Selector)
 	if err != nil {
@@ -395,43 +349,6 @@ func (r *HvpaReconciler) reconcileVpa(hvpa *autoscalingv1alpha1.Hvpa) (*vpa_api.
 	}
 
 	return vpa.Status.DeepCopy(), nil
-}
-
-func getHpaFromHvpa(hvpa *autoscalingv1alpha1.Hvpa) (*autoscaling.HorizontalPodAutoscaler, error) {
-	metadata := hvpa.Spec.Hpa.Template.ObjectMeta.DeepCopy()
-
-	labels := metadata.GetLabels()
-	if labels == nil || len(labels) == 0 {
-		// TODO: Could be done better as part of validation
-		return nil, fmt.Errorf("Need labels in HPA template")
-	}
-
-	if ownerRef := metadata.GetOwnerReferences(); len(ownerRef) != 0 {
-		// TODO: Could be done better as part of validation
-		return nil, fmt.Errorf("hpa template in hvpa object already has an owner reference")
-	}
-
-	if generateName := metadata.GetGenerateName(); len(generateName) != 0 {
-		log.V(3).Info("Warning", "Generate name provided in the hpa template will be ignored", generateName)
-	}
-
-	if name := metadata.GetName(); len(name) != 0 {
-		log.V(3).Info("Warning", "Name provided in the hpa template will be ignored", name)
-		metadata.SetName("")
-	}
-
-	metadata.SetGenerateName(hvpa.Name + "-")
-	metadata.SetNamespace(hvpa.Namespace)
-
-	return &autoscaling.HorizontalPodAutoscaler{
-		ObjectMeta: *metadata,
-		Spec: autoscaling.HorizontalPodAutoscalerSpec{
-			MaxReplicas:    hvpa.Spec.Hpa.Template.Spec.MaxReplicas,
-			MinReplicas:    hvpa.Spec.Hpa.Template.Spec.MinReplicas,
-			ScaleTargetRef: *hvpa.Spec.TargetRef.DeepCopy(),
-			Metrics:        hvpa.Spec.Hpa.Template.Spec.Metrics,
-		},
-	}, nil
 }
 
 func (r *HvpaReconciler) reconcileHpa(hvpa *autoscalingv1alpha1.Hvpa) (*autoscaling.HorizontalPodAutoscalerStatus, error) {
